Use Python int's for position functions

+++ dask_ndmeasure/__init__.py
@@ -393,7 +393,7 @@ def maximum_position(input, labels=None, index=None):
         index = index.flatten()
 
     indices = _utils._ravel_shape_indices(
-        input.shape, dtype=numpy.int64, chunks=input.chunks
+        input.shape, chunks=input.chunks
     )
 
     max_lbl = maximum(input, labels=labels, index=index)
@@ -421,7 +421,7 @@ def maximum_position(input, labels=None, index=None):
             axis=tuple(_pycompat.irange(index.ndim, max_lbl_indices.ndim))
         ),
         0
-    ).astype(numpy.int64)
+    ).astype(indices.dtype)
 
     max_pos_lbl = []
     max_1dpos_lbl_rem = max_1dpos_lbl
@@ -601,7 +601,7 @@ def minimum_position(input, labels=None, index=None):
         index = index.flatten()
 
     indices = _utils._ravel_shape_indices(
-        input.shape, dtype=numpy.int64, chunks=input.chunks
+        input.shape, chunks=input.chunks
     )
 
     min_lbl = minimum(input, labels=labels, index=index)
@@ -629,7 +629,7 @@ def minimum_position(input, labels=None, index=None):
             axis=tuple(_pycompat.irange(index.ndim, min_lbl_indices.ndim))
         ),
         0
-    ).astype(numpy.int64)
+    ).astype(indices.dtype)
 
     min_pos_lbl = []
     min_1dpos_lbl_rem = min_1dpos_lbl
